==> jupyter_notebooks/2_feature_engineering.ipynb <==
# 2. Feature engineering

### Objectives
Use technical analysis to add features to a dataset

### Inputs
`output/nb1_market_index_with_historical.feather`

### Outputs
`output/nb2_market_index_with_historical_and_features.feather`

### Motes/comments

---

import pandas as pd
from ta.volatility import BollingerBands
from ta.trend import MACD, SMAIndicator
from ta.momentum import RSIIndicator


### Load data

returns_df = pd.read_feather('../output/nb1_market_index_with_historical.feather')
returns_df.head()

### Feature engineering

# We’ll apply indicators to the Lithium_Index (avg weekly return)
# Alternatively you can apply them to each ticker separately

# Example: SMA and RSI on Lithium_Index
sma = SMAIndicator(close=returns_df["Lithium_Index"], window=4)
rsi = RSIIndicator(close=returns_df["Lithium_Index"], window=4)

returns_df["SMA_4"] = sma.sma_indicator()
returns_df["RSI_4"] = rsi.rsi()

# MACD
macd = MACD(close=returns_df["Lithium_Index"])
returns_df["MACD"] = macd.macd()
returns_df["MACD_Signal"] = macd.macd_signal()

# Bollinger Bands (for volatility)
bb = BollingerBands(close=returns_df["Lithium_Index"])
returns_df["BB_High"] = bb.bollinger_hband()
returns_df["BB_Low"] = bb.bollinger_lband()
returns_df["BB_Width"] = returns_df["BB_High"] - returns_df["BB_Low"]

returns_df.head()

* Now the feature set includes:
	* Has_Event
	* SMA_4, RSI_4
	* MACD, MACD_Signal
	* BB_Width (volatility)

*Possible - include returns of each individual stock too if necessary (later).*

returns_df.to_feather('../output/nb2_market_index_with_historical_and_features.feather')

returns_df.columns

type(returns_df.index)
